populate the gallery with instagram pictures
The old pictures have been removed and the gallery div now
uses the Instagram Graph API to get recent posts, and adds
only those with the #forkspit tag.

Another hashtag can be used. At this moment #forkspit is
the most consistent for finding tattoo pictures.

TODO: add a try catch block in case fetch fails, and in that case,
populate the gallery using the old pictures.